+++ src/OpenTelemetry/Trace/ActivityExporter.cs
@@ -13,11 +13,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // </copyright>
+
 using System;
-using System.Collections.Generic;
 using System.Diagnostics;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace OpenTelemetry.Trace
 {
@@ -27,19 +25,14 @@ namespace OpenTelemetry.Trace
     public enum ExportResult
     {
         /// <summary>
-        /// Batch is successfully exported.
+        /// Batch export succeeded.
         /// </summary>
         Success = 0,
 
         /// <summary>
-        /// Batch export failed. Caller must not retry.
+        /// Batch export failed.
         /// </summary>
-        FailedNotRetryable = 1,
-
-        /// <summary>
-        /// Batch export failed transiently. Caller should record error and may retry.
-        /// </summary>
-        FailedRetryable = 2,
+        Failure = 1,
     }
 
     /// <summary>
@@ -48,19 +41,18 @@ namespace OpenTelemetry.Trace
     public abstract class ActivityExporter : IDisposable
     {
         /// <summary>
-        /// Exports batch of activities asynchronously.
+        /// Export a batch of <see cref="Activity"/> objects.
         /// </summary>
-        /// <param name="batch">Batch of activities to export.</param>
-        /// <param name="cancellationToken">Cancellation token.</param>
+        /// <param name="batch">Batch of <see cref="Activity"/> objects to export.</param>
         /// <returns>Result of export.</returns>
-        public abstract Task<ExportResult> ExportAsync(IEnumerable<Activity> batch, CancellationToken cancellationToken);
+        public abstract ExportResult Export(in Batch<Activity> batch);
 
         /// <summary>
-        /// Shuts down exporter asynchronously.
+        /// Shuts down the exporter.
         /// </summary>
-        /// <param name="cancellationToken">Cancellation token.</param>
-        /// <returns>Returns <see cref="Task"/>.</returns>
-        public abstract Task ShutdownAsync(CancellationToken cancellationToken);
+        public virtual void Shutdown()
+        {
+        }
 
         /// <inheritdoc/>
         public void Dispose()