+++ test/OpenTelemetry.Tests/Trace/LinkTest.cs
@@ -16,6 +16,7 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Linq;
 using Xunit;
 
 namespace OpenTelemetry.Trace.Tests
@@ -24,7 +25,7 @@ namespace OpenTelemetry.Trace.Tests
     {
         private readonly IDictionary<string, object> attributesMap = new Dictionary<string, object>();
         private readonly SpanContext spanContext;
-        private readonly ActivityTagsCollection tags;
+        private readonly SpanAttributes tags;
 
         public LinkTest()
         {
@@ -34,7 +35,19 @@ namespace OpenTelemetry.Trace.Tests
             this.attributesMap.Add("MyAttributeKey1", 10L);
             this.attributesMap.Add("MyAttributeKey2", true);
             this.attributesMap.Add("MyAttributeKey3", 0.005);
-            this.tags = new ActivityTagsCollection(this.attributesMap);
+            this.attributesMap.Add("MyAttributeKey4", new long[] { 1, 2 });
+            this.attributesMap.Add("MyAttributeKey5", new string[] { "a", "b" });
+            this.attributesMap.Add("MyAttributeKey6", new bool[] { true, false });
+            this.attributesMap.Add("MyAttributeKey7", new double[] { 0.1, -0.1 });
+            this.tags = new SpanAttributes();
+            this.tags.Add("MyAttributeKey0", "MyStringAttribute");
+            this.tags.Add("MyAttributeKey1", 10L);
+            this.tags.Add("MyAttributeKey2", true);
+            this.tags.Add("MyAttributeKey3", 0.005);
+            this.tags.Add("MyAttributeKey4", new long[] { 1, 2 });
+            this.tags.Add("MyAttributeKey5", new string[] { "a", "b" });
+            this.tags.Add("MyAttributeKey6", new bool[] { true, false });
+            this.tags.Add("MyAttributeKey7", new double[] { 0.1, -0.1 });
         }
 
         [Fact]
@@ -51,7 +64,11 @@ namespace OpenTelemetry.Trace.Tests
             var link = new Link(this.spanContext, this.tags);
             Assert.Equal(this.spanContext.TraceId, link.Context.TraceId);
             Assert.Equal(this.spanContext.SpanId, link.Context.SpanId);
-            Assert.Equal(this.attributesMap, link.Attributes);
+
+            foreach (var attributemap in this.attributesMap)
+            {
+                Assert.Equal(attributemap.Value, link.Attributes.FirstOrDefault(a => a.Key == attributemap.Key).Value);
+            }
         }
 
         [Fact]
@@ -91,7 +108,7 @@ namespace OpenTelemetry.Trace.Tests
         [Fact]
         public void NotEquality_WithAttributes()
         {
-            var tag1 = new ActivityTagsCollection(new Dictionary<string, object>());
+            var tag1 = new SpanAttributes();
             var tag2 = this.tags;
             var link1 = new Link(this.spanContext, tag1);
             var link2 = new Link(this.spanContext, tag2);